Feature/correct

+++ src/main/kotlin/com/sparta/techTree/post/service/PostServiceImpl.kt
@@ -10,17 +10,19 @@ import com.sparta.techTree.post.model.Post
 import com.sparta.techTree.post.model.toResponse
 import com.sparta.techTree.post.repository.PostRepository
 import org.springframework.data.repository.findByIdOrNull
-
 import com.sparta.techTree.like.repository.LikeRepository
 import com.sparta.techTree.user.model.UserEntity
 import com.sparta.techTree.user.repository.UserRepository
-import jakarta.persistence.EntityNotFoundException
-import org.springframework.security.core.context.SecurityContextHolder
+import java.nio.file.AccessDeniedException
 
 
 @Service
 
-class PostServiceImpl(private val postRepository: PostRepository, private val likeRepository: LikeRepository,private val userRepository: UserRepository) :
+class PostServiceImpl(
+    private val postRepository: PostRepository,
+    private val likeRepository: LikeRepository,
+    private val userRepository: UserRepository
+) :
     PostService {
 
 
@@ -42,10 +44,7 @@ class PostServiceImpl(private val postRepository: PostRepository, private val li
 
 
     override fun createPost(request: CreatePostRequest, userId: Long): PostResponse {
-        val user = userRepository.findById(userId)
-            .orElseThrow {
-                EntityNotFoundException("User with ID $userId not found")
-            }
+        val user: UserEntity = userRepository.findByIdOrNull(userId) ?: throw ModelNotFoundException("User", userId)
         val createdPost = postRepository.save(
             Post(
                 title = request.title,
@@ -58,18 +57,27 @@ class PostServiceImpl(private val postRepository: PostRepository, private val li
     }
 
     @Transactional
-    override fun updatePost(postId: Long, request: UpdatePostRequest,): PostResponse {
+    override fun updatePost(postId: Long, userId: Long, request: UpdatePostRequest): PostResponse {
         val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
-        val countLikes = likeRepository.countByPostId(postId)
+        if (post.user.id != userId) {
+            throw AccessDeniedException("User with ID $userId does not have permission to update post with ID $postId")
+        }
         post.title = request.title ?: post.title
         post.content = request.content ?: post.content
+        val updatePost = postRepository.save(post)
+        val countLikes = likeRepository.countByPostId(postId)
         post.countLikes = countLikes
-        return post.toResponse()
+
+
+        return updatePost.toResponse()
     }
 
     @Transactional
-    override fun deletePost(postId: Long) {
+    override fun deletePost(postId: Long, userId: Long) {
         val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+        if (post.user.id != userId) {
+            throw AccessDeniedException("User with ID $userId does not have permission to update post with ID $postId")
+        }
         postRepository.delete(post)
     }
 }
